requirements.txt
ccxt
python-dotenv
numpy
pandas
scikit-learn
